[RENAME] entity 안으로 폴더 이동
